refactor: optimize BookService and BookProgressService

+++ backend/src/main/java/com/example/readflow/stats/dto/StatsOverviewDto.java
@@ -3,8 +3,8 @@ package com.example.readflow.stats.dto;
 import java.util.List;
 
 public record StatsOverviewDto(
-        int totalBooks,
-        int completedBooks,
+        long totalBooks,
+        long completedBooks,
         long totalPagesRead,
         long totalReadingMinutes,
         int currentStreak,
